refactor: move setCurrentUserShareSpacePermissions to helper
