+++ packages/web-pkg/src/helpers/spaces/index.ts
@@ -1,0 +1,1 @@
+export * from './permissions'
