+++ packages/web-pkg/src/services/folder/loaders/loaderSpace.ts
@@ -8,15 +8,18 @@ import {
   isPersonalSpaceResource,
   isPublicSpaceResource,
   isShareSpaceResource,
-  SpaceMember,
   SpaceResource
 } from '@opencloud-eu/web-client'
 import { unref } from 'vue'
 import { FolderLoaderOptions } from './types'
 import { DriveItem } from '@opencloud-eu/web-client/graph/generated'
 import { isLocationSpacesActive, isLocationPublicActive } from '../../../router'
-import { SharesStore, SpacesStore, useFileRouteReplace, UserStore } from '../../../composables'
-import { getIndicators, getSharedDriveItem } from '../../../helpers'
+import { useFileRouteReplace } from '../../../composables'
+import {
+  getIndicators,
+  getSharedDriveItem,
+  setCurrentUserShareSpacePermissions
+} from '../../../helpers'
 
 export class FolderLoaderSpace implements FolderLoader {
   public isEnabled(): boolean {
@@ -47,8 +50,6 @@ export class FolderLoaderSpace implements FolderLoader {
     } = context
     const { webdav, graphAuthenticated: graphClient } = clientService
     const { replaceInvalidFileRoute } = useFileRouteReplace({ router })
-
-    const setCurrentUserShareSpacePermissions = this.setCurrentUserShareSpacePermissions
 
     return useTask(function* (
       signal1,
@@ -140,54 +141,5 @@ export class FolderLoaderSpace implements FolderLoader {
         }
       }
     }).restartable()
-  }
-
-  /**
-   * Since shared spaces are only virtual, they and their permissions can't be fetched from the server.
-   * Hence the permissions for the current user need to be set manually via the corresponding drive item.
-   */
-  private setCurrentUserShareSpacePermissions({
-    sharesStore,
-    spacesStore,
-    userStore,
-    space,
-    sharedDriveItem
-  }: {
-    sharesStore: SharesStore
-    spacesStore: SpacesStore
-    userStore: UserStore
-    space: SpaceResource
-    sharedDriveItem: DriveItem
-  }) {
-    const permissions = sharedDriveItem?.remoteItem?.permissions || []
-    if (!permissions.length) {
-      return
-    }
-
-    const allPermissions: string[] = []
-    permissions.forEach((permission) => {
-      if (permission['@libre.graph.permissions.actions']) {
-        allPermissions.push(...permission['@libre.graph.permissions.actions'])
-        return
-      }
-      const role = sharesStore.graphRoles[permission.roles[0]]
-      if (!role) {
-        return
-      }
-      const permissions = role.rolePermissions.flatMap((p) => p.allowedResourceActions)
-      allPermissions.push(...permissions)
-    })
-
-    const uniquePermissions = [...new Set(allPermissions)]
-    const spaceMember: SpaceMember = {
-      grantedTo: { user: { id: userStore.user.id, displayName: userStore.user.displayName } },
-      permissions: uniquePermissions,
-      roleId: ''
-    }
-    spacesStore.updateSpaceField({
-      id: space.id,
-      field: 'members',
-      value: { [userStore.user.id]: spaceMember }
-    })
   }
 }
